requirements.txt
absl-py==1.0.0
astunparse==1.6.3
blis==0.7.5
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
cymem==2.0.6
dill==0.3.4
es-core-news-sm==3.2.0
et-xmlfile==1.1.0
filelock==3.4.2
flatbuffers==2.0
fonttools==4.29.0
gast==0.3.3
gensim==4.1.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
h5py==2.10.0
huggingface-hub==0.4.0
idna==3.3
imbalanced-learn==0.9.0
imblearn==0.0
importlib-metadata==4.10.1
importlib-resources==5.4.0
install==1.3.5
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
keras-nightly==2.9.0.dev2022021208
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langcodes==3.3.0
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
murmurhash==1.0.6
nltk==3.6.7
numpy==1.22.2
oauthlib==3.2.0
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pathy==0.6.1
Pillow==9.0.0
preshed==3.0.6
promise==2.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
sadice==0.1.3
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.11.2
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
tb-nightly==2.9.0a20220212
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-datasets==4.5.2
tensorflow-estimator==2.3.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-metadata==1.6.0
termcolor==1.1.0
tf-estimator-nightly==2.9.0.dev2022021209
tf-nightly-gpu==2.9.0.dev20220212
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchsampler==0.1.1
torchvision==0.11.3+cu113
tqdm==4.62.3
transformers==4.15.0
typer==0.4.0
typing-extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
Werkzeug==2.0.3
wrapt==1.13.3
xgboost==1.5.2
xlrd==1.2.0
zipp==3.7.0
